Long/Integer key bugfix

+++ src/me/pepe/DatabaseAPI/DatabaseManager/Types/Player/PlayerDatabaseTable.java
@@ -12,7 +12,7 @@ public abstract class PlayerDatabaseTable<V extends DatabaseTable> extends Datab
 	private PlayerData pData;
 	private long lastUpdate = 0;
 	public PlayerDatabaseTable(String name, PlayerData pData, Database database) {
-		super(name, "identifier", DatabaseKeyType.INT, database);
+		super(name, "identifier", DatabaseKeyType.LONG, database);
 		this.pData = pData;
 	}
 	@Override
